[Fix][Relax][Frontend][Torch] Emit int64 indices for sort and argsort
`torch.sort` and `torch.argsort` return int64 indices, but the Relax
frontend emitted int32. `relax.op.argsort` defaults to `dtype="int32"`,
and both call sites in `base_fx_graph_translator.py` took that default.

`_topk` in the same file already overrides the identical int32 default on
`relax.op.topk` to match PyTorch, and every other index-producing op the
frontend supports (argmax, argmin, max.dim, median.dim, bucketize) comes
out as int64. sort and argsort were the only exceptions, so a single
imported graph could carry two different index dtypes for the same kind
of value.

The sorted values and the index values themselves were already correct;
only the index dtype diverged.

Pass `dtype="int64"` at both call sites, update the expectations that had
the int32 result written into them, and add a `test_sort` to the
exported-program tests, which had no coverage for `torch.sort`.

+++ tests/python/relax/test_frontend_from_fx.py
@@ -6067,10 +6067,12 @@ def test_argsort():
         @R.function
         def main(
             inp_0: R.Tensor((5, 3), dtype="float32"),
-        ) -> R.Tensor((5, 3), dtype="int32"):
+        ) -> R.Tensor((5, 3), dtype="int64"):
             with R.dataflow():
-                lv: R.Tensor((5, 3), dtype="int32") = R.argsort(inp_0, axis=1, descending=True)
-                gv: R.Tensor((5, 3), dtype="int32") = lv
+                lv: R.Tensor((5, 3), dtype="int64") = R.argsort(
+                    inp_0, axis=1, descending=True, dtype="int64"
+                )
+                gv: R.Tensor((5, 3), dtype="int64") = lv
                 R.output(gv)
             return gv
 
@@ -6086,18 +6088,18 @@ def test_sort():
     class Expected:
         @R.function
         def main(inp_0: R.Tensor((5, 3), dtype="float32")) -> R.Tuple(
-            R.Tensor((5, 3), dtype="float32"), R.Tensor((5, 3), dtype="int32")
+            R.Tensor((5, 3), dtype="float32"), R.Tensor((5, 3), dtype="int64")
         ):
             with R.dataflow():
-                lv: R.Tensor((5, 3), dtype="int32") = R.argsort(
-                    inp_0, axis=1, descending=True, dtype="int32"
+                lv: R.Tensor((5, 3), dtype="int64") = R.argsort(
+                    inp_0, axis=1, descending=True, dtype="int64"
                 )
                 lv1: R.Tensor((5, 3), dtype="float32") = R.gather_elements(inp_0, lv, axis=1)
-                lv2: R.Tuple(R.Tensor((5, 3), dtype="float32"), R.Tensor((5, 3), dtype="int32")) = (
+                lv2: R.Tuple(R.Tensor((5, 3), dtype="float32"), R.Tensor((5, 3), dtype="int64")) = (
                     lv1,
                     lv,
                 )
-                gv: R.Tuple(R.Tensor((5, 3), dtype="float32"), R.Tensor((5, 3), dtype="int32")) = (
+                gv: R.Tuple(R.Tensor((5, 3), dtype="float32"), R.Tensor((5, 3), dtype="int64")) = (
                     lv2
                 )
                 R.output(gv)
